Add swaps app utils
